Fixed bug in parallel example..
The bug was that I forgot to delete jobs of numbers that were not
multiples of 3 or 5. This made jobs accumulate in tube 'A' over time.
It had no effect on the end result, but multiple runs would take
more and more time as 'A' would have a lot of garbage.

Also changed drain_all example to output the number of drained jobs,
but not the jobs themselves.